add new dao

+++ iscript-service/iscript-service-crawler/src/main/java/com/lezo/iscript/service/crawler/service/ProxyAddrService.java
@@ -11,5 +11,5 @@ public interface ProxyAddrService {
 
 	ProxyAddrDto getProxyAddrDto(String ipString, int port);
 	
-	ProxyAddrDto getProxyAddrDto(int ipValue, int port);
+	ProxyAddrDto getProxyAddrDto(Long ipValue, int port);
 }
